exercises update dynamically through json, types file with the blooms taxonomy labels for the levels to provide actual learning experience
