+++ my-app/src/data/interval-exercises.json
@@ -1,0 +1,36 @@
+[
+  {
+    "id": "chord-1",
+    "name": "Major vs Minor (C)",
+    "difficulty": "beginner",
+    "category": "chords",
+    "unlockLevel": 1,
+    "steps": [
+      {
+        "level": "Remember",
+        "type": "multipleChoice",
+        "question": "Which interval is greater?",
+        "audio": "C4",
+        "options": ["C", "D"],
+        "answer": "C"
+      }
+    ]
+  },
+  {
+    "id": "chord-2",
+    "name": "Chord Progressions I",
+    "difficulty": "beginner",
+    "category": "chords",
+    "unlockLevel": 2,
+    "steps": [
+      {
+        "level": "Apply",
+        "type": "multipleChoice",
+        "question": "What chord follows C in a I-IV-V progression?",
+        "audio": "F4",
+        "options": ["Fmaj", "Gmaj", "Amin"],
+        "answer": "Fmaj"
+      }
+    ]
+  }
+]
