+++ my-app/src/data/intervalExercises.ts
@@ -1,0 +1,4 @@
+import data from "./interval-exercises.json";
+import type { Exercise } from "./types";
+
+export const intervalExercises = data as Exercise[];
