+++ my-app/src/data/noteExercises.ts
@@ -1,0 +1,4 @@
+import data from "./note-exercises.json";
+import type { Exercise } from "./types";
+
+export const noteExercises = data as Exercise[];
